web/requirements.txt
git+https:///github.com/maxisme/idmyteam-server.git#subdirectory=web/idmyteam
django==3.1.2
oyaml==1.0
Pillow==8.0.1
websocket-client==0.57.0
opencv-python==4.4.0.44
requests==2.24.0
